+++ pkl-core/src/main/java/org/pkl/core/ast/expression/member/AbstractInvokeQualifiedMethodNode.java
@@ -15,37 +15,29 @@
  */
 package org.pkl.core.ast.expression.member;
 
-import com.oracle.truffle.api.CallTarget;
+import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.source.SourceSection;
 import org.pkl.core.ast.ExpressionNode;
-import org.pkl.core.ast.VmModifier;
 import org.pkl.core.runtime.Identifier;
-import org.pkl.core.runtime.VmObjectLike;
+import org.pkl.core.runtime.VmTyped;
 
-/** A non-virtual call of an object method that is lexically scoped. */
-public final class InvokeLexicalObjectMethodNode extends AbstractInvokeLexicalMethodNode {
-  public InvokeLexicalObjectMethodNode(
+public abstract sealed class AbstractInvokeQualifiedMethodNode extends AbstractInvokeMethodNode
+    permits InvokeQualifiedClassMethodNode, InvokeQualifiedObjectMethodNode {
+  @Child private ExpressionNode getReceiverNode;
+
+  protected AbstractInvokeQualifiedMethodNode(
       SourceSection sourceSection,
       Identifier methodName,
-      int levelsUp,
       ExpressionNode[] argumentNodes,
-      boolean needsConst) {
-    super(sourceSection, methodName, levelsUp, argumentNodes, needsConst);
+      boolean needsConst,
+      ExpressionNode getReceiverNode) {
+    super(sourceSection, methodName, argumentNodes, needsConst);
+    this.getReceiverNode = getReceiverNode;
   }
 
   @Override
-  protected void doCheckConst(VmObjectLike owner) {
-    var member = owner.getMember(methodName);
-    assert member != null;
-    if (!VmModifier.isConst(member.getModifiers())) {
-      throw exceptionBuilder().evalError("methodMustBeConst", methodName).build();
-    }
-  }
-
-  @Override
-  protected CallTarget getCallTarget(VmObjectLike owner) {
-    var method = owner.getMember(methodName);
-    assert method != null && method.isLocal();
-    return (CallTarget) method.getCallTarget().call(owner, owner);
+  public final Object executeGeneric(VirtualFrame frame) {
+    var receiver = (VmTyped) getReceiverNode.executeGeneric(frame);
+    return invoke(frame, receiver, receiver);
   }
 }
