+++ pkl-core/src/main/java/org/pkl/core/ast/expression/member/InvokeQualifiedObjectMethodNode.java
@@ -16,32 +16,21 @@
 package org.pkl.core.ast.expression.member;
 
 import com.oracle.truffle.api.CallTarget;
-import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.source.SourceSection;
 import org.pkl.core.ast.ExpressionNode;
 import org.pkl.core.ast.VmModifier;
 import org.pkl.core.runtime.Identifier;
 import org.pkl.core.runtime.VmObjectLike;
-import org.pkl.core.runtime.VmTyped;
 
 /** A non-virtual call of an object method, invoked off of an explicit receiver. */
-public final class InvokeQualifiedObjectMethodNode extends AbstractInvokeMethodNode {
-  @Child private ExpressionNode getReceiverNode;
-
+public final class InvokeQualifiedObjectMethodNode extends AbstractInvokeQualifiedMethodNode {
   public InvokeQualifiedObjectMethodNode(
       SourceSection sourceSection,
       Identifier methodName,
       ExpressionNode[] argumentNodes,
       boolean needsConst,
       ExpressionNode getReceiverNode) {
-    super(sourceSection, methodName, argumentNodes, needsConst);
-    this.getReceiverNode = getReceiverNode;
-  }
-
-  @Override
-  public Object executeGeneric(VirtualFrame frame) {
-    var receiver = (VmTyped) getReceiverNode.executeGeneric(frame);
-    return invoke(frame, receiver, receiver);
+    super(sourceSection, methodName, argumentNodes, needsConst, getReceiverNode);
   }
 
   @Override
