+++ pkl-core/src/test/files/LanguageSnippetTests/input/types/typeAliasConstraint6.pkl
