+++ pkl-core/src/test/files/LanguageSnippetTests/input/types/typeAliasConstraint7.pkl
